[FIX] [#1] 모임 생성 request 데이터 바인딩 안되는 오류 수정
